Plural name for the controller of vehicle type
